requirements.txt
Flask>=2.3.0
flask-cors>=3.1.0
Werkzeug>=2.3.0
Pillow>=10.0.0
numpy>=1.26.0
scikit-image>=0.21.0
